Removed -a option
Basically it forces all executions of this program to run like "-a" was set to true, in other words, to always send requests to every url, regardless his cname. This allows to add new services in fingerprints.go without the need to put a cname where "nxdomain" field is false (i.e. when the takeover identification depends only on the http response, and not if a nxdomain cname exists).